--- app/categoria/audifonos/Audifonos/index.tsx
@@ -1,1 +1,0 @@
-export { default } from "./Audifonos";
