--- app/categoria/audifonos/[pId]/ProductoDetalles/index.tsx
@@ -1,1 +1,0 @@
-export { default } from './ProductoDetalles';
