--- app/categoria/teclados/[pId]/ProductoDetalles/index.tsx
@@ -1,1 +1,0 @@
-export { default } from "./ProductoDetalles";
